fix(cookies bug): fix the bug
