+++ src/middleware.js
@@ -1,3 +1,5 @@
+import { createServerClient } from "@supabase/ssr";
+import { NextResponse } from "next/server";
 import createMiddleware from "next-intl/middleware";
 
 export default createMiddleware({
@@ -13,3 +15,61 @@ export const config = {
     // certain folders and all pathnames with a dot (e.g. favicon.ico)
     matcher: ["/((?!api|_next|_vercel|.*\\..*).*)"],
 };
+
+export async function middleware(request) {
+    let response = NextResponse.next({
+        request: {
+            headers: request.headers,
+        },
+    });
+
+    const supabase = createServerClient(
+        process.env.NEXT_PUBLIC_SUPABASE_URL,
+        process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
+        {
+            cookies: {
+                get(name) {
+                    return request.cookies.get(name)?.value;
+                },
+                set(name, value, options) {
+                    request.cookies.set({
+                        name,
+                        value,
+                        ...options,
+                    });
+                    response = NextResponse.next({
+                        request: {
+                            headers: request.headers,
+                        },
+                    });
+                    response.cookies.set({
+                        name,
+                        value,
+                        ...options,
+                    });
+                },
+                remove(name, options) {
+                    request.cookies.set({
+                        name,
+                        value: "",
+                        ...options,
+                    });
+                    response = NextResponse.next({
+                        request: {
+                            headers: request.headers,
+                        },
+                    });
+                    response.cookies.set({
+                        name,
+                        value: "",
+                        ...options,
+                    });
+                },
+            },
+        }
+    );
+
+    await supabase.auth.getSession();
+
+    return response;
+}
